move getLookupDict
